% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contained_in_tex.R
\name{minibib}
\alias{minibib}
\title{Return a list of references is contained in a set of tex files and write that list to a new bibtex file.}
\usage{
minibib(bibtex, write_bib = TRUE, path = ".", pattern = "\\\\.tex$",
  recursive = TRUE, file = "main.bib", ...)
}
\arguments{
\item{bibtex}{character path to bibtex file}

\item{write_bib}{a logical value. If TRUE, writes a bibtex file with the references contained in the tex file.}

\item{path}{character vector of full path names}

\item{pattern}{a regular expression. File names matching the regular expression are returned.}

\item{recursive}{a logical value. Should subdirectories be searched for tex files as well?}

\item{...}{further arguments passed to write.bib in the bibtex package}
}
\description{
Return a list of references is contained in a set of tex files and write that list to a new bibtex file.
}
